Move TermColor enum into pretty formater
